relocate logging dir & update createLogger imports
