+++ src/api/common/helpers/request-logger.js
@@ -1,6 +1,6 @@
 import hapiPino from 'hapi-pino'
 
-import { loggerOptions } from './logger-options.js'
+import { loggerOptions } from '../../../logging/logger-options.js'
 
 const requestLogger = {
   plugin: hapiPino,
