+++ src/api/common/helpers/start-server.js
@@ -1,6 +1,6 @@
 import { config } from '../../../config/index.js'
 import { createServer } from '../../index.js'
-import { createLogger } from './logging/logger.js'
+import { createLogger } from '../../../logging/logger.js'
 
 const startServer = async () => {
   let server
